JADESView Update for JWST images
I've fixed it so that now this works with actual JWST images, and I've made it so that the user doesn't have to specify BEAGLE or EAZY plots for the program to run. In addition, I've added support for using a segmentation map plot, providing you use SEGMAP as the filter name in the image list.